Reader: Switch to ExternalLink, add icon to timestamp on full post.
This patch switches a number of external links in the reader to use the ExternalLink component. In one case, it does not (the timestamp in PostByline) because the default ExternalLink forces an icon size of 18, which doesn't work in the 14px byline.

We could add a size param to ExternalLink, but this patch is simpler without it.