Domain Management: loosen restrictions for resource records.
As per RFC 2181, there's actually only one restriction for DNS records - length.
But realistically speaking, we should do some sane checking.
For A/AAAA records we enforce a valid *hostname* as per RFC 952 and RFC 1123.
For other resource records, we are more forgiving - we check the length (can be empty, max 255)
and allow only certain characters that are expected in such records, but in any order.